google_adk/backend/requirements.txt
google-adk[otel-gcp]
litellm
httpx
python-dotenv
uvicorn
fastapi
pydantic
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-httpx 
opentelemetry-instrumentation-system-metrics 
openinference-instrumentation-google-adk
opentelemetry-instrumentation-threading
openinference-instrumentation-mcp
opentelemetry-instrumentation-fastapi
openinference-instrumentation-litellm
openinference-instrumentation-anthropic
openinference-instrumentation-openai
opentelemetry-instrumentation-aiohttp-client